+++ backend/src/main/java/com/sjiwon/anotherart/global/security/dto/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.sjiwon.anotherart.global.security;
+package com.sjiwon.anotherart.global.security.dto;
 
 public record LoginRequest(
         String id,
